refactor: avoid unnecessary asdf GitHub API calls
Previously, we would check for the latest version of an asdf plugin.
However, this often incurs an API call to GitHub. There's a real
possibility of reaching the unauthenticated API rate limit for GitHub,
so this means that even if a plugin were already installed, the setup
playbook would fail because the latest version check would always run.

Instead, we're checking first that the plugin doesn't exist yet, and
only then allowing a latest version check. Also, grouping those tasks
into a block so we can treat them cohesively.