refactor: rebuild the tmux config
Removing most of the old tmux config which was complex and difficult to
configure. Colors always felt a bit off, and there was way more in there
than I needed. Moving to a simpler config with options that I understand
and put in there from scratch, and now suddenly colors look better (such
as with the "catppuccin" theme.